ntdll: Better match the PSDK types and fix the winapi_check warnings.

+++ dlls/ntdll/string.c
@@ -105,8 +105,8 @@ void __cdecl NTDLL_qsort( void *base, size_t nmemb, size_t size,
 /*********************************************************************
  *                  _lfind   (NTDLL.@)
  */
-void * __cdecl _lfind( const void *key, const void *base, size_t *nmemb,
-                       size_t size, int(*compar)(const void *, const void *) )
+void * __cdecl _lfind( const void *key, const void *base, unsigned int *nmemb,
+                       unsigned int size, int(*compar)(const void *, const void *) )
 {
     return lfind( key, base, nmemb, size, compar );
 }
